refactor: decompose web.py into Flask blueprints
Splits ~2,000-line web.py into 7 blueprints + shared helpers.
Zero behavior change — all routes respond identically.

- web.py: app factory only (184 lines, down from ~2,000)
- blueprints/auth.py: /login, /logout (seed for future LDAP/OIDC)
- blueprints/audit.py: /audit, /bulk_audit, /diff, /connect
- blueprints/history.py: /archive/*, /activity/*
- blueprints/schedules.py: /schedules/*
- blueprints/settings_bp.py: /settings/*, /license/*
- blueprints/reports.py: /reports/*
- blueprints/api_v1.py: /api/v1/*
- _vendor_helpers.py: detect_vendor(), validate_vendor_format(), extract_hostname()
- _helpers.py: _require_auth_impl(), _err(), _make_temp_path()

All 181 tests pass. ruff clean.
