+++ src/cashel/_helpers.py
@@ -1,0 +1,92 @@
+"""Shared non-route utility helpers for Cashel.
+
+Includes _require_auth_impl() so web.py stays lean while keeping
+_require_auth registered as app.before_request (per task spec).
+"""
+
+import logging as _logging
+import os
+import secrets
+import tempfile
+import time
+import uuid
+
+from flask import g, jsonify, redirect, request, session, url_for
+
+from .settings import get_settings
+
+_logger = _logging.getLogger(__name__)
+
+_MAX_FILE_BYTES = 5 * 1024 * 1024  # 5 MB per-file limit enforced in routes
+
+UPLOAD_FOLDER = os.environ.get("UPLOAD_FOLDER", "/tmp/cashel_uploads")
+
+
+def _make_temp_path(suffix: str) -> str:
+    """Return a writable temp path, preferring UPLOAD_FOLDER with system-temp fallback."""
+    candidate = os.path.join(UPLOAD_FOLDER, f"{uuid.uuid4()}{suffix}")
+    try:
+        fd = os.open(candidate, os.O_CREAT | os.O_WRONLY, 0o600)
+        os.close(fd)
+        os.unlink(candidate)
+        return candidate
+    except OSError:
+        fd, path = tempfile.mkstemp(suffix=suffix)
+        os.close(fd)
+        return path
+
+
+_AUTH_EXEMPT_ENDPOINTS = {"auth.login", "auth.login_post", "auth.logout", "health", "static"}
+
+
+def _require_auth_impl(demo_mode: bool):
+    """Core auth gate — called by web.py's app.before_request hook."""
+    if demo_mode:
+        g.auth_method = "demo"
+        return
+
+    settings = get_settings()
+    if not settings.get("auth_enabled"):
+        return
+
+    if request.endpoint in _AUTH_EXEMPT_ENDPOINTS:
+        return
+
+    # API key auth (header or query param) — used by CLI/CI clients
+    api_key_header = request.headers.get("X-API-Key") or request.args.get("api_key")
+    if api_key_header:
+        stored = settings.get("api_key", "")
+        if stored and secrets.compare_digest(api_key_header, stored):
+            g.auth_method = "api_key"
+            return
+        if request.path.startswith("/api/"):
+            return jsonify({"ok": False, "data": None, "error": "Invalid API key."}), 401
+        return jsonify({"error": "Invalid API key."}), 401
+
+    # Session auth (browser)
+    if session.get("authenticated"):
+        lifetime = settings.get("session_lifetime_minutes", 480)
+        if time.time() - session.get("last_seen", 0) < lifetime * 60:
+            session["last_seen"] = time.time()
+            g.auth_method = "session"
+            return
+        session.clear()
+
+    # Not authenticated
+    if request.path.startswith("/api/"):
+        return jsonify({"ok": False, "data": None, "error": "Authentication required."}), 401
+    next_url = request.url if request.method == "GET" else None
+    return redirect(url_for("auth.login", next=next_url))
+
+
+def _err(exc: Exception, generic_msg: str = "An internal error occurred.") -> str:
+    """Return an error message string respecting the error_detail setting.
+
+    In 'full' mode (development) the raw exception is surfaced.
+    In 'sanitized' mode (production default) only *generic_msg* is returned and
+    the full exception is written to the application log.
+    """
+    _logger.exception("Internal error: %s", exc)
+    if get_settings().get("error_detail") == "full":
+        return str(exc)
+    return generic_msg
